chore: add missing React keys and rename user constants to grade
