+++ src/hooks/queries/useMe.ts
@@ -1,7 +1,7 @@
 import { useSuspenseQuery } from '@tanstack/react-query';
 import { http } from '../../utils/http';
 import { queryKeys } from './queryKeys';
-import type { Grade } from '@/domains/user';
+import type { Grade } from '@/domains/grade';
 
 interface GetMeResponse {
   point: number;
